월드 정보 수정

+++ mua/util/scheduler.py
@@ -3,22 +3,25 @@ from mua.models import World
 from mua import db
 
 
-def updateWorld(app):
+def updateWorld(app, url):
     """
     월드 이름과 종류를 업데이트 하기 위한 함수
     """
-    WORLD_URL = "https://maplestory.nexon.com/N23Ranking/World/Total"
     with app.app_context():
-        response = getWorldInfo(WORLD_URL)
+        response = getWorldInfo(url)
         normal_world: dict = response.get("일반 월드")
         for world in normal_world:
-            world_model = World(name=world, type="normal")
-            db.session.add(world_model)
+            world_name = World.query.get(world)
+            if not world_name:
+                world_model = World(name=world, type="normal")
+                db.session.add(world_model)
 
         reboot_world: dict = response.get("리부트 월드")
         for world in reboot_world:
-            world_model = World(name=world, type="reboot")
-            db.session.add(world_model)
+            world_name = World.query.get(world)
+            if not world_name:
+                world_model = World(name=world, type="reboot")
+                db.session.add(world_model)
         db.session.commit()
         db.session.close()
